pyproject.toml
[tool.poetry]
name = "quick_batch"
version = "0.1.0"
description = ""
authors = ["Jeremy Watt <jermwatt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
docker = "^6.1.2"
fire = "^0.5.0"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
flake8 = "^4.0.1"
mypy = "^0.910"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
